refactor : Exception to RunTimeException 변경
